+++ src/frontend/apps/e2e/__tests__/app-impress/doc-comments.spec.ts
@@ -6,7 +6,7 @@ import {
   getOtherBrowserName,
   verifyDocName,
 } from './utils-common';
-import { writeInEditor } from './utils-editor';
+import { getEditor, writeInEditor } from './utils-editor';
 import {
   addNewMember,
   connectOtherUserToDoc,
@@ -48,6 +48,7 @@ test.describe('Doc Comments', () => {
     await thread.locator('[data-test="save"]').click();
     await expect(thread.getByText('This is a comment').first()).toBeHidden();
 
+    await editor.first().click();
     await editor.getByText('Hello').click();
 
     await thread.getByText('This is a comment').first().hover();
@@ -135,6 +136,7 @@ test.describe('Doc Comments', () => {
       'background-color',
       'rgba(237, 180, 0, 0.4)',
     );
+    await editor.first().click();
     await editor.getByText('Hello').click();
 
     await thread.getByText('This is a comment').first().hover();
@@ -197,6 +199,7 @@ test.describe('Doc Comments', () => {
       'background-color',
       'rgba(237, 180, 0, 0.4)',
     );
+    await editor.first().click();
     await editor.getByText('Hello').click();
 
     await thread.getByText('This is a new comment').first().hover();
@@ -216,6 +219,8 @@ test.describe('Doc Comments', () => {
 
     // We share the doc with another user
     const otherBrowserName = getOtherBrowserName(browserName);
+
+    const editor = await getEditor({ page });
 
     // Add a new member with editor role
     await page.getByRole('button', { name: 'Share' }).click();
@@ -240,7 +245,7 @@ test.describe('Doc Comments', () => {
       text: 'Hello, I can edit the document',
     });
     await expect(
-      otherEditor.getByText('Hello, I can edit the document'),
+      editor.getByText('Hello, I can edit the document'),
     ).toBeVisible();
     await otherEditor.getByText('Hello').selectText();
     await otherPage.getByRole('button', { name: 'Comment' }).click();
@@ -317,6 +322,49 @@ test.describe('Doc Comments', () => {
     await expect(otherThread.locator('[data-test="moreactions"]')).toBeHidden();
 
     await cleanup();
+  });
+
+  test('it checks comments pasting from another document', async ({
+    page,
+    browserName,
+  }) => {
+    await createDoc(page, 'comment-doc-1', browserName, 1);
+
+    // We add a comment in the first document
+    const editor1 = await writeInEditor({ page, text: 'Document One' });
+    await editor1.getByText('Document One').selectText();
+    await page.getByRole('button', { name: 'Comment' }).click();
+
+    const thread1 = page.locator('.bn-thread');
+    await thread1.getByRole('paragraph').first().fill('Comment in Doc One');
+    await thread1.locator('[data-test="save"]').click();
+    await expect(thread1.getByText('Comment in Doc One').first()).toBeHidden();
+
+    await expect(editor1.getByText('Document One')).toHaveCSS(
+      'background-color',
+      'rgba(237, 180, 0, 0.4)',
+    );
+
+    await editor1.getByText('Document One').click();
+    // We copy the content including the comment from the first document
+    await editor1.getByText('Document One').selectText();
+    await page.keyboard.press('Control+C');
+
+    // We create a second document
+    await createDoc(page, 'comment-doc-2', browserName, 1);
+
+    // We paste the content into the second document
+    const editor2 = await writeInEditor({ page, text: '' });
+    await editor2.click();
+    await page.keyboard.press('Control+V');
+
+    await expect(editor2.getByText('Document One')).toHaveCSS(
+      'background-color',
+      'rgba(0, 0, 0, 0)',
+    );
+
+    await editor2.getByText('Document One').click();
+    await expect(page.locator('.bn-thread')).toBeHidden();
   });
 });