+++ src/frontend/apps/e2e/__tests__/app-impress/doc-version.spec.ts
@@ -6,6 +6,7 @@ import {
   mockedDocument,
   verifyDocName,
 } from './utils-common';
+import { openSuggestionMenu, writeInEditor } from './utils-editor';
 
 test.beforeEach(async ({ page }) => {
   await page.goto('/');
@@ -17,6 +18,7 @@ test.describe('Doc Version', () => {
 
     await verifyDocName(page, randomDoc);
 
+    // Initially, there is no version
     await page.getByLabel('Open the document options').click();
     await page.getByRole('menuitem', { name: 'Version history' }).click();
     await expect(page.getByText('History', { exact: true })).toBeVisible();
@@ -26,31 +28,49 @@ test.describe('Doc Version', () => {
     await expect(panel).toBeVisible();
     await expect(modal.getByText('No versions')).toBeVisible();
 
-    const editor = page.locator('.ProseMirror');
     await modal.getByRole('button', { name: 'close' }).click();
-    await editor.click();
-    await page.keyboard.type('# Hello World');
 
+    await writeInEditor({ page, text: 'Hello World' });
+
+    // It will trigger a save, no version created yet (initial version is not counted)
     await goToGridDoc(page, {
       title: randomDoc,
     });
 
-    await expect(
-      page.getByRole('heading', { name: 'Hello World' }),
-    ).toBeVisible();
+    await expect(page.getByText('Hello World')).toBeVisible();
 
-    await page
-      .locator('.ProseMirror .bn-block')
-      .getByRole('heading', { name: 'Hello World' })
-      .fill('It will create a version');
+    // Write more
+    await writeInEditor({ page, text: 'It will create a version' });
 
+    await openSuggestionMenu({ page });
+    await page.getByText('Add a callout block').click();
+
+    const calloutBlock = page
+      .locator('div[data-content-type="callout"]')
+      .first();
+
+    await expect(calloutBlock).toBeVisible();
+
+    // It will trigger a save and create a version this time
     await goToGridDoc(page, {
       title: randomDoc,
     });
 
     await expect(page.getByText('Hello World')).toBeHidden();
+    await expect(page.getByText('It will create a version')).toBeVisible();
+
+    await expect(calloutBlock).toBeVisible();
+
+    // Write more
+    await writeInEditor({ page, text: 'It will create a second version' });
+
+    // It will trigger a save and create a second version
+    await goToGridDoc(page, {
+      title: randomDoc,
+    });
+
     await expect(
-      page.getByRole('heading', { name: 'It will create a version' }),
+      page.getByText('It will create a second version'),
     ).toBeVisible();
 
     await page.getByLabel('Open the document options').click();
@@ -60,11 +80,33 @@ test.describe('Doc Version', () => {
     await expect(page.getByText('History', { exact: true })).toBeVisible();
     await expect(page.getByRole('status')).toBeHidden();
     const items = await panel.locator('.version-item').all();
-    expect(items.length).toBe(1);
-    await items[0].click();
+    expect(items.length).toBe(2);
+    await items[1].click();
 
     await expect(modal.getByText('Hello World')).toBeVisible();
     await expect(modal.getByText('It will create a version')).toBeHidden();
+    await expect(
+      modal.locator('div[data-content-type="callout"]').first(),
+    ).toBeHidden();
+
+    await items[0].click();
+
+    await expect(modal.getByText('Hello World')).toBeVisible();
+    await expect(modal.getByText('It will create a version')).toBeVisible();
+    await expect(
+      modal.locator('div[data-content-type="callout"]').first(),
+    ).toBeVisible();
+    await expect(
+      modal.getByText('It will create a second version'),
+    ).toBeHidden();
+
+    await items[1].click();
+
+    await expect(modal.getByText('Hello World')).toBeVisible();
+    await expect(modal.getByText('It will create a version')).toBeHidden();
+    await expect(
+      modal.locator('div[data-content-type="callout"]').first(),
+    ).toBeHidden();
   });
 
   test('it does not display the doc versions if not allowed', async ({