+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/DocEditor.tsx
@@ -90,7 +90,7 @@ export const DocEditor = ({ doc }: DocEditorProps) => {
     }
   }, [isProviderReady, setIsSkeletonVisible]);
 
-  if (!isProviderReady) {
+  if (!isProviderReady || provider?.configuration.name !== doc.id) {
     return <Loading />;
   }