+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/custom-blocks/CalloutBlock.tsx
@@ -5,8 +5,8 @@ import {
   InlineContentSchema,
   StyleSchema,
   defaultProps,
-  insertOrUpdateBlock,
 } from '@blocknote/core';
+import { insertOrUpdateBlockForSlashMenu } from '@blocknote/core/extensions';
 import { BlockTypeSelectItem, createReactBlockSpec } from '@blocknote/react';
 import { TFunction } from 'i18next';
 import React, { useEffect, useState } from 'react';
@@ -97,34 +97,41 @@ const CalloutComponent = ({
       `}
     >
       <CalloutBlockStyle />
-      <BoxButton
-        contentEditable={false}
-        onClick={toggleEmojiPicker}
-        $css={css`
-          font-size: 1.125rem;
-          cursor: ${isEditable ? 'pointer' : 'default'};
-          ${isEditable &&
-          `
+      <Box $position="relative">
+        <BoxButton
+          contentEditable={false}
+          onClick={toggleEmojiPicker}
+          onKeyDown={(e) => {
+            if (e.key === 'Escape' && openEmojiPicker) {
+              setOpenEmojiPicker(false);
+            }
+          }}
+          $css={css`
+            font-size: 1.125rem;
+            cursor: ${isEditable ? 'pointer' : 'default'};
+            ${isEditable &&
+            `
           &:hover {
             background-color: rgba(0, 0, 0, 0.1);
           }
           `}
-        `}
-        $align="center"
-        $width="28px"
-        $radius="4px"
-      >
-        {block.props.emoji}
-      </BoxButton>
+          `}
+          $align="center"
+          $width="28px"
+          $radius="4px"
+        >
+          {block.props.emoji}
+        </BoxButton>
 
-      {openEmojiPicker && (
-        <EmojiPicker
-          emojiData={emojidata}
-          onClickOutside={onClickOutside}
-          onEmojiSelect={onEmojiSelect}
-          withOverlay={true}
-        />
-      )}
+        {openEmojiPicker && (
+          <EmojiPicker
+            emojiData={emojidata}
+            onClickOutside={onClickOutside}
+            onEmojiSelect={onEmojiSelect}
+            withOverlay={true}
+          />
+        )}
+      </Box>
       <Box as="p" className="inline-content" ref={contentRef} />
     </Box>
   );
@@ -156,7 +163,7 @@ export const getCalloutReactSlashMenuItems = (
     key: 'callout',
     title: t('Callout'),
     onItemClick: () => {
-      insertOrUpdateBlock(editor, {
+      insertOrUpdateBlockForSlashMenu(editor, {
         type: 'callout',
       });
     },