+++ src/frontend/apps/impress/src/features/docs/doc-versioning/api/useDocVersions.tsx
@@ -77,7 +77,6 @@ export function useDocVersionsInfiniteQuery(
     getNextPageParam(lastPage) {
       return lastPage.next_version_id_marker || undefined;
     },
-
     ...queryConfig,
   });
 }