+++ CHANGELOG.md
@@ -6,15 +6,31 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [4.2.0] - 2025-12-17
+
 ### Added
 
-- ✨(backend) allow to create a new user in a marketing system
+- ✨(backend) allow to create a new user in a marketing system #1707
+- ✨(backend) add async indexation of documents on save (or access save) #1276
+- ✨(backend) add debounce mechanism to limit indexation jobs #1276
+- ✨(api) add API route to search for indexed documents in Find #1276
+- 🥅(frontend) add boundary error page #1728
 - ✨(backend) manage reconciliation requests for user accounts #1708
 
 ### Changed
 
+- 🛂(backend) stop throttling collaboration servers #1730
 - 🚸(backend) use unaccented full name for user search #1637
 - 🌐(backend) internationalize demo #1644
+- ♿(frontend) improve accessibility:
+  - ♿️Improve keyboard accessibility for the document tree #1681
+
+### Fixed
+
+- 🐛(frontend) paste content with comments from another document #1732
+- 🐛(frontend) Select text + Go back one page crash the app #1733
+- 🐛(frontend) fix versioning conflict #1742
+
 
 ## [4.1.0] - 2025-12-09
 
@@ -929,7 +945,8 @@ and this project adheres to
 - ✨(frontend) Coming Soon page (#67)
 - 🚀 Impress, project to manage your documents easily and collaboratively.
 
-[unreleased]: https://github.com/suitenumerique/docs/compare/v4.1.0...main
+[unreleased]: https://github.com/suitenumerique/docs/compare/v4.2.0...main
+[v4.2.0]: https://github.com/suitenumerique/docs/releases/v4.2.0
 [v4.1.0]: https://github.com/suitenumerique/docs/releases/v4.1.0
 [v4.0.0]: https://github.com/suitenumerique/docs/releases/v4.0.0
 [v3.10.0]: https://github.com/suitenumerique/docs/releases/v3.10.0